Update naming/wording across project.
I didn't want to limit perception of the project to simply storing different
versions of personal projects or applications, as it could be used for much more
given the right mindset. To this end, I've changed some of the language to be
much more general and hopefully inspire a few more uses.